chore(qa): resolve SonarQube findings
Unnamed _ for unused catch/lambda bindings, drop unthrowable checked
exceptions from test signatures, rename record-named variables, tighten
assertion chains, plus a few small simplifications.

on-behalf-of: @camptocamp <info@camptocamp.com>

+++ integrations/parquetry-tileverse-storage/src/main/java/io/tileverse/parquetry/tileverse/RangeReaderByteRangeSource.java
@@ -68,7 +68,7 @@ final class RangeReaderByteRangeSource implements ByteRangeSource {
         int requested = (int) length;
         int total = 0;
         while (total < requested) {
-            ByteBuffer target = dst.asSlice(total, requested - total).asByteBuffer();
+            ByteBuffer target = dst.asSlice(total, (long) requested - total).asByteBuffer();
             int read = reader.readRange(offset + total, requested - total, target);
             if (read <= 0) {
                 break;
